requirements.txt
fastapi
uvicorn
pandas
numpy
faiss-cpu
openai
scikit-learn
sentence-transformers
python-multipart
streamlit  # (optional)
